+++ spec/ruby/core/array/intersect_spec.rb
@@ -4,13 +4,15 @@ describe 'Array#intersect?' do
   ruby_version_is '3.1' do # https://bugs.ruby-lang.org/issues/15198
     describe 'when at least one element in two Arrays is the same' do
       it 'returns true' do
-        [1, 2].intersect?([2, 3]).should == true
+        [1, 2].intersect?([2, 3, 4]).should == true
+        [2, 3, 4].intersect?([1, 2]).should == true
       end
     end
 
     describe 'when there are no elements in common between two Arrays' do
       it 'returns false' do
-        [1, 2].intersect?([3, 4]).should == false
+        [0, 1, 2].intersect?([3, 4]).should == false
+        [3, 4].intersect?([0, 1, 2]).should == false
       end
     end
   end